Users

+++ server/controllers/thoughts.js
@@ -13,13 +13,13 @@ export const getThoughts = async (req, res) => {
 }
 
 export const createThought = async (req, res) => {
-    const newThought = new ThoughtModel(req.body)
+    const thought = req.body
+    const newThought = new ThoughtModel({ ...thought, author: { ...thought.author, _id: req.userId }})
 
     try{
         await newThought.save()
 
         res.status(201).json(newThought)
-        res.redirect
     } catch (error) {
         res.status(409).json({ message: error.message })
     }
@@ -28,21 +28,37 @@ export const createThought = async (req, res) => {
 export const updateThought = async (req, res) => {
     const { id } = req.params
 
-    if (!mongoose.Types.ObjectId.isValid(id)) return res.status(404).send(`No post with id: ${id}`)
+    try {
+        if (!mongoose.Types.ObjectId.isValid(id)) return res.status(404).send(`No post with id: ${id}`)
 
-    const updatedThought = { ...req.body, _id: id }
+        const thought = ThoughtModel.findById(id)
+        if (thought.author._id !== req.userId) return res.status(401).json({ message: "Unauthorized" })
+        const updatedThought = { ...req.body, _id: id }
 
-    await ThoughtModel.findByIdAndUpdate(id, updatedThought, { new: true })
+        await ThoughtModel.findByIdAndUpdate(id, updatedThought, { new: true })
 
-    res.json(updatedThought)
+        res.json(updatedThought)
+    } catch (error) {
+        res.status(500).json({ message: 'Something went wrong.'})
+    }
 }
 
 export const deleteThought = async (req, res) => {
     const { id } = req.params
 
-    if (!mongoose.Types.ObjectId.isValid(id)) return res.status(404).send(`No post with id: ${id}`)
+    try {
+        if (!mongoose.Types.ObjectId.isValid(id)) return res.status(404).send(`No post with id: ${id}`)
     
-    await ThoughtModel.findByIdAndDelete(id)
+        const thought = await ThoughtModel.findById(id)
 
-    res.json({ message: 'Deleted Successfully' })
+        if (thought.author._id !== req.userId) {
+            return res.status(401).json({ message: "Unauthorized" })
+        }
+
+        await ThoughtModel.findByIdAndDelete(id)
+
+        res.json({ message: 'Deleted Successfully' })
+    } catch (error) {
+        res.status(500).json({ message: 'Something went wrong.'})
+    }
 }
